My check for whether a class was nested or not was entirely broken -- it left out the common case that the class is being used inside the outer class without a qualifier.  A new test shows where we would do the wrong thing in that case.
Fixing this brought up more problems, that could all be traced
to the fact we 'cheat' and create fake uses of
forward-declares when we want to be sure not to delete them.
I get rid of the cheating and just say, when we see a
forward-decalare, whether we want to definitely-keep it or
not.  That yields much nicer code (I think).

I also don't know what I was thinking when I said "if there
are multiple fwd-decls in a class, just keep one arbitrarily."
Obviously we should keep the first one; the arbitrary one may
come too late.  These changes brought that bug to light too
(an existing test started failing), but now it's squashed.

Finally, I resolved a TODO, getting rid of
IsDeclNodeInsideFriend.  I had known this was possible for a
while, I just had never gotten around to actually doing it.
Doesn't affect any tests.

R=dsturtevant
DELTA=141  (75 added, 26 deleted, 40 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=1453