Correctly deal with non-class types that can be redeclared: functions, typedefs, etc.  We just require one of the redecls to be visible, not all of them.
This fix isn't perfect -- we just check for redecls in the
same file as the use, and don't do all the complicated stuff
we do to pick the 'best' #include for a redecl if necessary,
like we do for forward-declarations.  (The situation is
different, since we are willing to suggest adding a new
forward-decl for a class, but not for these other types like
functions.)  But it's still better than it was before.

Submitting TBR since it's the weekend, and this is a
pre-requisite for getting the big typedef fix in.


Revision created by MOE tool push_codebase.
MOE_MIGRATION=1600