iwyu was egregiously wrong in how it handled template arguments using the 'precomputed cache'.  In such situations, it totally ignored the currently active resugar_map, replacing it with one of its own.  That worked fine for types outside of templates, but not fine for types inside (such as a 'hash_map<T>' inside a templated class).
I "fixed" this.  "Fixed" is in quotes because this turned up a
whole slew of other problems I don't even attempt to resolve
here (though I spent a few hours trying).  One is that it's
possible to have a type like hash_map that has some arguments
that are dependent and some that aren't; in theory, for these
types, we can correctly attribute the use to the template
author or template instantiator depending on which type it
is.  But I can't figure out how to get clang to do any
meaningful analysis of incomplete (dependent) types, so I've
punted on that for now.

The second thing wrong is I jumped through all sorts of hoops
to handle default template arguments correctly, so if a class
has a hash_map<T> and you instantiate T with string, you're
also made responsible for hash<string>.  This *should* work,
but clang is giving hash<string> a type I don't expect
(RecordType, not TemplateSpecializationType), and I don't know
how to deal with that -- I don't know how to extract the
'string' part of this RecordType.  Ugh.  I punt on this now,
as well.

Even in this incomplete form, it's enough to resolve a P1 bug,
so I figure it's worth putting in.

R=dsturtevant
DELTA=141  (97 added, 7 deleted, 37 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3147