Add an IWYU test that a crashing bug in Clang doesn't regress.
If Sema::DefineImplicitDefaultConstructor is called for a
template class with an integer template argument, it used to
create a corresponding TemplateTypeLoc of Expression type but
with NULL Expression field. Subsequently, if getSourceRange()
is called on TemplateTypeLoc, it would SEGV due to trying to
dereference the return of getSourceExpression().

The fix is to properly inside a TemplateTypeLoc of Expression type.

This CL comprised IWYU test which crashed before the Clang fix.

R=matthewbg
DELTA=22  (22 added, 0 deleted, 0 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3894