[iwyu_tool] Add --basedir argument
Enable specifiying a base dir relative to which the IWYU output should
be interpreted.

With the path normalization that happens when interpreting IWYU output,
we now need to normalize command-line arguments as well, in order for
filtering to work correctly.

Let test framework simulate a cwd to be able to test this.