Ignore friend decls when looking for definition-as-written
Clang r283207 changed how redeclaration chains are built (we're not yet
sure if this was intentional) so that friend declarations inside
instantiated templates are broken off the normal redecl chain for a
function. E.g.

  // template.h
  void foo();

  template<class T>
  struct Template {
    friend void foo();
  };

  // foo.h
  inline void foo() {
  }

  // main.cpp
  #include "template.h"
  #include "foo.h"

  Template<int> i;

  int main() {
    foo();  // attributed to Template<T>::foo
  }

If Template<int> is instantiated, uses of foo can see Template<int>::foo
as the primary declaration instead of the inline function in foo.h.

And there's no longer a redeclaration chain from Template<T>::foo to the
other foo's, so IWYU can't find the 'real' definition-as-written.

This patch works around the problem by ignoring friend declarations in
templates when looking for a declaration-as-written.