Remove incorrect test case assumption on desugared types
Clang r289250 caused more type sugar to remain in the AST. That, in turn, caused
derived_function_tpl_args to fail, because we had a compensating assertion
(marked with a TODO indicating that it was undesirable.)

Now that type sugar survives, we won't see a need for IndirectClass
here. Rather, the use will be attributed to the LocalClass typedef, which exists
in the same file. No diagnostic expected.