Exclude prefix headers from macro-defined-by-includer heuristics (issue #368).
Also considered options of not populating `macro_users_` or
`direct_includes_as_fileentries_` for prefix headers but decided not to
do so. I think it is better to have correct information for prefix
headers but not to give recommendations based on that instead of having
incomplete information.

PR https://github.com/include-what-you-use/include-what-you-use/pull/410