Fix requiring allocator for getting container iterator with libc++. (#448)
libc++ code is

    template <class _Tp, class _Allocator>
    inline _LIBCPP_INLINE_VISIBILITY
    typename vector<_Tp, _Allocator>::iterator
    vector<_Tp, _Allocator>::begin() _NOEXCEPT
    {
        return __make_iter(this->__begin_);
    }

and we reported use of allocator when visiting `vector<_Tp, _Allocator>`
because here `_Allocator` is not a default template argument. There was
no such problem in libstdc++ because there `begin` is defined inside a class.

Also I wasn't able to reproduce the issue with non-STL types because for
not precomputed types we replayed template arguments usage. And when we
populate the cache we see which template arguments are default so there
are no problems during replay.

Update test to remove method calls using object template arguments. As result,
on macOS it fixed errors

    tests/cxx/badinc.cc:1534: Unexpected diagnostic:
    I2_Enum is defined in "tests/cxx/badinc-i2.h", which isn't directly #included.

    tests/cxx/badinc.cc:1580: Unexpected diagnostic:
    I2_Enum is defined in "tests/cxx/badinc-i2.h", which isn't directly #included.

but introduced another error

    tests/cxx/badinc.cc:1527: Unmatched regex:
    I2_Enum is...*badinc-i2.h

I think it's a reasonable trade-off.