Enhance fix_includes.py to understand Allman namespaces
The fix_includes.py script was originally written to work for namespaces
with { on the same line as the namespace. For styles that do not follow
this format, namespace detection does not work properly as the line
numbers and regex statement are not able to be evaluated.

Add logic to detect Allman namespaces and shift line numbers
appropriately, so that the following namespace formatting can be
handled:

  namespace A
  {
    class Foo;
  }

Also detect mixed same-line and Allman formatting, e.g.

  namespace A { namespace B { namespace C
  {
    class Foo;
  }
  }
  }

With the addition of Allman namespace detection, mixed namespace
detection has also been added to cover all namespace styles.

Signed-off-by: Christian Venegas <cvenegas@esri.com>