Set CMake policy CMP0048 unconditionally
Now that we require CMake 3.4.3, we no longer need to check if CMP0048
is available before setting it to NEW.

Also move the policy setting and project declaration so we only create a
project for out-of-tree builds. In-tree builds will become part of the
LLVM project, which already has this policy set.