[iwyu_tool] Use extra args verbatim
This is a breaking change.

Originally, the extra arguments after '--' were only intended for IWYU
args, and so '-Xiwyu' was automatically prepended to each of them to
make life easier on users.

But this made it impossible to add additional Clang args, which is also
a useful feature.

The extra arguments are now forwarded as-written to main and users need
to manually add '-Xiwyu' before every IWYU argument.