Support headers using symbols mapped to themselves
Currently if IWYU decides that the best header for a symbol is the same
one that uses the symbol, it can cause that header to add a #include of
itself.

This in particular can happen with mappings.

Detect this case and cause the use to be ignored.

Furthermore, when choosing from multiple public mapping headers, prefer
to use oneself where possible (since that minimizes #includes).