Don't crash if using decl comes from PCH
If a using-declaration is in a precompiled header, we won't have any
file info registered for it. Motivating use case:

  // main.cpp
  int main() {
    A a;
  }

  // pch.h
  struct A {};
  using ::A;

  $ clang++ -x c++-header pch.h -o pch.h.pch
  $ include-what-you-use.exe -c -include pch.h main.cpp

Work around this by skipping using declarations without an associated
file, but assert that this only happens if we have an implicit PCH on
the command-line.

We don't have infrastructure for testing with actual precompiled
headers, so no test case.

Originally PR #688 by Zachary Henkel, squashed by Kim Gräsman.