Add skipped associated headers to IWYU analysis
This was detected by an assertion in iwyu_output.h:

  Assertion failed: desired_includes_have_been_calculated_ &&
      "Must calculate desired includes before calling
       desired_includes()"

This came about because there are two parallel structures for headers:

* IwyuFileInfo: direct includes, as quoted include strings
* IwyuPreprocessorInfo: files to report iwyu violations for, as
  FileEntry pointers

The assertion fires if a header ends up in IwyuFileInfo but not in
files-to-report, and iff that header is an associated header.

Therefore, make sure we call AddGlobToReportIWYUViolationsFor in all
places where we classify something as an associated header.

A minimal repro case is:

  // tests/cxx/x.cc
  #include "y.h"
  #include "tests/cxx/x.h"
  #include "x.h"

if include-what-you-use is invoked with the absolute path of x.cc.

Our current test infrastructure doesn't support absolute paths for the
test target, so this can't be reproduced. Add a test to exercise the
problematic path even if it doesn't trigger the assertion.

Root-causing and original fix by Nick Overdijk, tests and rationale by
Kim Gräsman.

Fixes issue #738.