Prevent deleting duplicate nested forward decls
fix_includes.py could be over-zealous and delete forward declarations of
nested classes if they happened to share a name with another nested
class in the same file.  This was because the analysis of top-level
spans detected #ifdefs and namespaces, but not classes as things which
made code non-top-level.

Luckily, the IWYU output makes it obvious when a forward declaration is
for a nested class.  Take advantage of that to fix this bug.

Also, add a regression test.