Fix another alias-in-template issue
It wasn't correct to ignore all uses of a template argument somewhere
below a TypedefNameDecl ASTNode.  When a template specialization happens
between the two, the typedef is irrelevant.  Account for that, and add a
suitable test case.

This makes cases like this count as a full use of the argument T:

template<typename T>
struct Container {
  using value_type = pair<T>;
};

which may not always be correct, but is sometimes necessary.  Currently
this is erring on the side of too many full-uses.