Simplify implicit ctor/dtor instantiation
This is a rewrite of the earlier instantiation logic, which was
unnecessarily general and complicated.

Rather than enumerating all constructors, use
Sema::LookupDefaultConstructor to get the one we're interested in.

Looking at assertions and expectations in the Lookup and Define methods
in Sema, it was possible to further simplify conditions to match, e.g.

* Cannot perform lookup on a decl that lacks definition, is in a
  dependent context or is being defined (the last was previously omitted)
* Three things are required to define any implicit method: it must be
  defaulted, it must not be deleted and it cannot have a body

This makes the logic much easier to understand while preserving the
current behavior.

(It should now be trivial to extend with e.g. implicit assignment or
other implicitly-generated methods.)